Revert "Added S3"
This reverts commit 275dfb2374adf43e12c7085e29465b17ce1d248f.

+++ src/main/java/com/project/TabernasSevilla/domain/Review.java
@@ -28,7 +28,7 @@ public class Review extends BaseEntity{
     private String comment;
  
     @Column(name = "rating", nullable = true)
-    private Double rating;
+    private int rating;
     
     /*
     @ManyToOne
